refactor(client): widget-layer foundation for the list-shaped settings tabs (UI/UX v3 P1c prep) (#52)
* refactor(client): widget-layer foundation for the list-shaped settings tabs

Preparation for migrating Ssh / Keybindings / Profiles onto the widget
layer. All three share one shape — a list, an edit panel for the selected
entry, Add/Delete buttons, and a delete dialog — and none of it is
expressible today.

Split `draw.rs` by control family. It sat at 779 lines against the
project's 800-line guideline, so the new control kinds would not have fit.
`draw/mod.rs` keeps the `draw_widget` dispatch and the chrome every control
shares (row fill, label column, focus ring, text baseline); `controls.rs`,
`slider.rs`, `text.rs` and `list.rs` own one family each. No behaviour
change: the moved code is byte-identical apart from visibility.

Widen `WidgetId.index` from u8 to u16. The migrated tabs are fixed-length
(14 rows at most), but a list-shaped category addresses one widget per
list entry and those lists are user-populated with no upper bound — a u8
would have stopped addressing entries past the 256th with no diagnostic.
`as_u32` shifts by 16 instead of 8; the widest offset is 0xFF_FFFF, well
inside the 100M-wide node-id slot reserved for SETTINGS_WIDGET_BASE. The
five per-tab focus counters that hold widget indices follow. Ssh's and
Keybindings' counters keep their own numbering until those tabs migrate.

Add the three missing `WidgetKind` arms — `ListItem`, `Button` and
`KeyCapture` — with their drawing and their AccessKit roles
(ListBoxOption / Button / TextInput). A key capture is not text entry: it
swallows the next key press rather than appending it, and a reader has to
announce that difference, so it does not reuse `Text`.

Give `Text` a real caret. The renderer appended `_` to the value and
ignored `TextInputState.cursor` entirely, so the caret sat at the end of
the string however the user had moved it — this already affected the
migrated Security byte-cap fields, the Startup shell fields and the font
fallback list. `caret` is a byte offset that falls back to end-of-string
when absent or mid-character.

Tests: 1651 pass (+15). IME preedit is still not shown in these fields;
that is pre-existing and left for the tab migrations.

via [Happy](https://happy.engineering)


* style(client): restore edition-2024 import ordering in accessibility.rs

A local editor hook reformatted the file with bare rustfmt, which defaults
to the 2015 style edition and sorts uppercase items last. The workspace is
on edition 2024, where cargo fmt sorts them first, so CI's fmt check failed
on all three platforms. Formatting only; no logic change.

via [Happy](https://happy.engineering)


---------

+++ nexterm-client-gpu/src/renderer/event_handler/mouse.rs
@@ -375,7 +375,7 @@ impl EventHandler {
             use crate::renderer::overlay::widgets::settings_window::WINDOW_CATEGORY;
             let hovered = match hit {
                 SettingsPanelHit::ThemeColor(i) => {
-                    Some((THEME_CATEGORY, THEME_SWATCH_BASE + i as u8))
+                    Some((THEME_CATEGORY, THEME_SWATCH_BASE + i as u16))
                 }
                 SettingsPanelHit::ThemeRow(index) => Some((THEME_CATEGORY, index)),
                 SettingsPanelHit::WindowRow(index) => Some((WINDOW_CATEGORY, index)),
